requirements.txt
flask
pyserial
pymysql